+++ hospital_traversing_robot/Programming/stc/mpu6050.h
@@ -156,7 +156,7 @@ int16_t get_raw_accel(uint8_t ind);
 int16_t get_raw_gyro(uint8_t ind);
 int32_t get_accel(uint8_t ind);
 int32_t get_gyro(uint8_t ind);
-int8_t get_accel_roll(void);
-int8_t get_accel_pitch(void);
+int16_t get_accel_roll(void);
+int16_t get_accel_pitch(void);
 
 #endif
